requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.1
chromadb==0.4.15
pandas>=2.2.0
openpyxl==3.1.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
numpy<2.0
python-dotenv==1.0.0